requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
pymssql
pyodbc
python-dotenv
bcrypt
numpy
pandas
scikit-learn
tensorflow
Pillow
python-dateutil
openai
langchain
langchain-openai
google-generativeai
requests